进一步完善以及修复BUG

+++ src/components/TashDetail.vue
@@ -49,6 +49,14 @@ export default {
         this.getNotebooks();
         this.getTrashNotes().then(() => {
             this.setCurTrashNote({ curTrashNoteId: this.$route.query.noteId });
+            if (!this.$route.query.noteId) {
+                this.$router.replace({
+                    path: '/trash',
+                    query: {
+                        noteId: this.curTrashNote.id,
+                    },
+                });
+            }
         });
     },
     computed: {
@@ -68,10 +76,30 @@ export default {
         ]),
         ...mapMutations(['setCurTrashNote']),
         onDelete() {
-            this.deleteTrashNote({ noteId: this.curTrashNote.id });
+            this.$confirm('删除后将无法恢复', '确定删除？', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning',
+            })
+                .then(() => {
+                    return this.deleteTrashNote({ noteId: this.curTrashNote.id });
+                })
+                .then(() => {
+                    console.log('delete success');
+                    this.setCurTrashNote({});
+                    this.$router.replace({
+                        path: '/trash',
+                        query: { noteId: this.curTrashNote.id },
+                    });
+                });
         },
         onRevert() {
-            this.revertTrashNote({ noteId: this.curTrashNote.id });
+            this.revertTrashNote({ noteId: this.curTrashNote.id }).then(() => {});
+            this.setCurTrashNote({ curTrashNoteId: {} });
+            this.$router.replace({
+                path: '/trash',
+                query: { noteId: this.curTrashNote.id },
+            });
         },
     },
     beforeRouteUpdate(to, from, next) {
